# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.1.0",
    "title": "dolphin-rs",
    "description": "A simple data logging server written in Rust.",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1"
    }
  ],
  "paths": {
    "/streams": {
      "get": {
        "summary": "List streams",
        "operationId": "listStreams",
        "tags": [
          "streams"
        ],
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Stream"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/streams/{key}/data": {
      "get": {
        "summary": "List logged data for stream",
        "operationId": "listData",
        "tags": [
          "data"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "description": "The key of the stream to retrieve",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Data"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add data to a stream",
        "operationId": "addData",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "description": "The key of the stream to retrieve",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "tags": [
          "data"
        ],
        "responses": {
          "200": {
            "description": "Null response"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Stream": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "key": {
            "type": "string"
          },
          "created": {
            "type": "string"
          },
          "updated": {
            "type": "string"
          }
        }
      },
      "Data": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "payload": {
            "type": "object"
          },
          "created": {
            "type": "string"
          },
          "updated": {
            "type": "string"
          }
        }
      }
    }
  }
}